Use usize for UnsafeLayoutNode.

+++ components/layout/wrapper.rs
@@ -1135,11 +1135,11 @@ pub trait PostorderNodeMutTraversal {
 
 /// Opaque type stored in type-unsafe work queues for parallel layout.
 /// Must be transmutable to and from LayoutNode/ThreadSafeLayoutNode.
-pub type UnsafeLayoutNode = (uint, uint);
+pub type UnsafeLayoutNode = (usize, usize);
 
 pub fn layout_node_to_unsafe_layout_node(node: &LayoutNode) -> UnsafeLayoutNode {
     unsafe {
-        let ptr: uint = mem::transmute_copy(node);
+        let ptr: usize = mem::transmute_copy(node);
         (ptr, 0)
     }
 }
